requirements.txt
fastapi~=0.112.0
uvicorn~=0.30.5
motor~=3.5.1
pydantic~=2.8.2
python-dotenv~=1.0.1
starlette~=0.37.2
shortuuid~=1.0.13
pytest~=8.3.2
httpx~=0.27.0
pytest-asyncio~=0.23.8
apscheduler